Update OVM_L1CrossDomainMessenger (#76)
* Update OVM_L1CrossDomainMessenger

Fixes problem in the value we were verifying.

* Fix broken tests

+++ packages/contracts/contracts/optimistic-ethereum/OVM/bridge/OVM_L1CrossDomainMessenger.sol
@@ -237,7 +237,7 @@ contract OVM_L1CrossDomainMessenger is iOVM_L1CrossDomainMessenger, OVM_BaseCros
 
         return Lib_SecureMerkleTrie.verifyInclusionProof(
             abi.encodePacked(storageKey),
-            abi.encodePacked(uint256(1)),
+            abi.encodePacked(uint8(1)),
             _proof.storageTrieWitness,
             account.storageRoot
         );
